return type changed to Double

+++ src/main/java/paymentStrategy/CashPaymentStrategy.java
@@ -4,7 +4,7 @@ import model.PaymentStatus;
 
 public class CashPaymentStrategy implements PaymentStrategy {
     @Override
-    public PaymentStatus makePayment(Long parkingFee) {
+    public PaymentStatus makePayment(Double parkingFee) {
         System.out.println("Payment of Rs. " + parkingFee + " done through Cash");
         return PaymentStatus.SUCCESS;
     }
